// src/app/web-client/components/call-center-call-view/enums.ts
export enum StateName {
    OffHook,
    Monitoring,
    FirstAgentCallRinging,
    WarmTransferInviteRinging,
    ColdTransferInviteRinging,
    WarmInviteAcceptedFirstAgent,
    WarmInviteAcceptedSecondAgent,
    FirstNormalAgentConnected,
    FirstOperatorAgentConnected
}

export enum ConferenceStatusAction {

    Accepted = 0,
    CallerClosedBeforeAnswer = 1,
    CallerClosedAfterAnswer = 2,
    SupervisorClosed = 3,
    AgentClosed = 4,
    OnHold = 5,
    Terminated = 6,
    Ringing = 7,
    Ignored = 8,
    Resumed = 9,
    Monitoring = 10,
    PickedUp = 11,
    WarmInviteSent = 12,
    WarmInviteAccepted = 13,
    WarmInviteRejected = 14,
    WarmAccepted = 15,
    WarmCanceled = 16,
    WarmSwitchToCaller = 17,
    WarmSwitchToAgent = 18,
    OperatorPark = 19,
    OperatorHold = 20,
    OperatorColdTransfered = 21,
    OperatorColdTransferInviteSent = 22,
    OperatorColdTransferInviteAccepted = 23,
    OperatorColdTransferInviteRejected = 24,
    OperatorColdTransferInviteCancelled = 25,
    Fallback = 26,
    SimultaniousIgnored = 27,
    OperatorInterruptEnded = 28,
    CallerAudioAddedToConference = 29,
    TrustedParticipantAudioAdded = 30,
    TrustedParticipantForAutoRecordingAdded = 31,
    TrustedParticipantForManualRecordingAdded = 32,
    TrustedParticipantForVoiceMailRecordingAdded = 33,
    TrustedParticipantAudioForAgentsAdded = 34,
    ConferenceReceivedFromPool = 35,
    NewConferenceCreated = 36,
    PersonalParked = 37,
    MonitoringEnded = 38,

}